+++ src/Storage/ExprCalc.Storage.Tests/DatabaseManagement/SqliteDbControllerTests.cs
@@ -1,4 +1,5 @@
 ﻿using ExprCalc.Entities;
+using ExprCalc.Entities.MetadataParams;
 using ExprCalc.Storage.Resources.DatabaseManagement;
 using ExprCalc.Storage.Resources.SqliteQueries;
 using ExprCalc.Storage.Resources.SqliteQueries.Exceptions;
@@ -29,20 +30,28 @@ namespace ExprCalc.Storage.Tests.DatabaseManagement
         {
             return DateTime.UnixEpoch + TimeSpan.FromMilliseconds((long)(dateTime - DateTime.UnixEpoch).TotalMilliseconds);
         }
-        private static Calculation CreateCalculation(string expression = "1 + 2", CalculationStatus? status = null, TimeSpan? timeOffset = null)
+        private static Calculation CreateCalculation(string expression = "1 + 2", CalculationStatus? status = null, string? user = null, TimeSpan? timeOffset = null)
         {
             DateTime createdAt = DateTime.UtcNow;
             if (timeOffset != null)
                 createdAt = createdAt.Add(timeOffset.Value);
             createdAt = RoundTime(createdAt);
 
-            return new Calculation(Guid.CreateVersion7(), expression, new User("test_user"), createdAt, createdAt, status ?? CalculationStatus.Pending);
+            return new Calculation(Guid.CreateVersion7(), expression, new User(user ?? "test_user"), createdAt, createdAt, status ?? CalculationStatus.Pending);
         }
-        private static Calculation CreateCalculationSuccess(string expression = "1 + 2", double result = 3, TimeSpan? timeOffset = null)
+        private static Calculation CreateCalculationPending(string expression = "1 + 2", string? user = null, TimeSpan? timeOffset = null)
         {
-            return CreateCalculation(expression, CalculationStatus.CreateSuccess(result), timeOffset);
+            return CreateCalculation(expression, CalculationStatus.Pending, user, timeOffset);
         }
-        private static Calculation CreateCalculationFailed(string expression = "1 + 2", TimeSpan? timeOffset = null)
+        private static Calculation CreateCalculationInProgress(string expression = "1 + 2", string? user = null, TimeSpan? timeOffset = null)
+        {
+            return CreateCalculation(expression, CalculationStatus.InProgress, user, timeOffset);
+        }
+        private static Calculation CreateCalculationSuccess(string expression = "1 + 2", string? user = null, double result = 3, TimeSpan? timeOffset = null)
+        {
+            return CreateCalculation(expression, CalculationStatus.CreateSuccess(result), user, timeOffset);
+        }
+        private static Calculation CreateCalculationFailed(string expression = "1 + 2", string? user = null, TimeSpan? timeOffset = null)
         {
             return CreateCalculation(expression,
                 CalculationStatus.CreateFailed(
@@ -52,12 +61,26 @@ namespace ExprCalc.Storage.Tests.DatabaseManagement
                         ErrorCode = CalculationErrorDetails.DivisionByZeroErrorCode,
                         Offset = 10,
                         Length = 1
-                    }), timeOffset);
+                    }), user, timeOffset);
 
         }
-        private static Calculation CreateCalculationCancelled(string expression = "1 + 2", string cancelledBy = "test_user2", TimeSpan? timeOffset = null)
+        private static Calculation CreateCalculationCancelled(string expression = "1 + 2", string? user = null, string cancelledBy = "test_user2", TimeSpan? timeOffset = null)
         {
-            return CreateCalculation(expression, CalculationStatus.CreateCancelled(new User(cancelledBy)), timeOffset);
+            return CreateCalculation(expression, CalculationStatus.CreateCancelled(new User(cancelledBy)), user, timeOffset);
+        }
+
+        private static IEnumerable<Calculation> ApplyFilter(IEnumerable<Calculation> calc, CalculationFilters filter)
+        {
+            return calc.Where(o =>
+                (filter.State == null || o.Status.State == filter.State.Value) &&
+                (filter.Id == null || o.Id == filter.Id.Value) &&
+                (filter.CreatedAtMin == null || o.CreatedAt >= filter.CreatedAtMin) &&
+                (filter.CreatedAtMax == null || o.CreatedAt < filter.CreatedAtMax) &&
+                (filter.UpdatedAtMin == null || o.UpdatedAt >= filter.UpdatedAtMin.Value) &&
+                (filter.UpdatedAtMax == null || o.UpdatedAt < filter.UpdatedAtMax.Value) &&
+                (filter.CreatedBy == null || o.CreatedBy.Login == filter.CreatedBy) &&
+                (filter.Expression == null || o.Expression.Contains(filter.Expression))
+                );
         }
 
         // ===============
@@ -143,7 +166,8 @@ namespace ExprCalc.Storage.Tests.DatabaseManagement
             await controller.AddCalculationAsync(insertionList[3], deadlockProtection.Token);
 
 
-            var list = await controller.GetCalculationsListAsync(deadlockProtection.Token);
+            var paginatedList = await controller.GetCalculationsListAsync(CalculationFilters.NoFilters, PaginationParams.AllData, deadlockProtection.Token);
+            var list = paginatedList.Items;
             list.Reverse();
 
             Assert.All(insertionList.Zip(list), (item) =>
@@ -153,6 +177,90 @@ namespace ExprCalc.Storage.Tests.DatabaseManagement
 
                 Assert.Equivalent(item.First, item.Second);
             });
+        }
+
+
+        [Fact]
+        public async Task GetCalculationsListWithFiltersTest()
+        {
+            CancellationTokenSource deadlockProtection = new CancellationTokenSource();
+            deadlockProtection.CancelAfter(TimeSpan.FromSeconds(60));
+
+            using var tempDirHolder = new TempDirectoryHolder();
+            await using var controller = CreateController(tempDirHolder.Directory);
+            await controller.Init(deadlockProtection.Token);
+
+            List<Calculation> insertionList = [
+                CreateCalculationPending(user: "abc"),
+                CreateCalculationInProgress(user: "bcd", timeOffset: TimeSpan.FromMilliseconds(10)),
+                CreateCalculationSuccess(user: "bcd", timeOffset: TimeSpan.FromMilliseconds(20)),
+                CreateCalculationFailed(user: "abc", expression: "log(10) / cos(1)", timeOffset: TimeSpan.FromMilliseconds(30)),
+                CreateCalculationCancelled(user: "abc", timeOffset: TimeSpan.FromMilliseconds(40))
+            ];
+
+            foreach (var item in insertionList)
+            {
+                await controller.AddCalculationAsync(item, deadlockProtection.Token);
+            }
+
+
+            List<CalculationFilters> filters = [
+                new CalculationFilters() { State = Entities.Enums.CalculationState.InProgress },
+                new CalculationFilters() { CreatedBy = "abc" },
+                new CalculationFilters() { Expression = "cos" },
+                new CalculationFilters() { Id = insertionList[0].Id },
+                new CalculationFilters() { CreatedAtMin = insertionList[1].CreatedAt, CreatedAtMax = insertionList[2].CreatedAt.AddMilliseconds(2) },
+                new CalculationFilters() { UpdatedAtMin = insertionList[1].CreatedAt, UpdatedAtMax = insertionList[2].CreatedAt.AddMilliseconds(2) },
+                new CalculationFilters() { UpdatedAtMin = insertionList[1].CreatedAt, CreatedBy = "abc" },
+            ];
+
+            foreach (var filter in filters)
+            {
+                var paginatedList = await controller.GetCalculationsListAsync(filter, PaginationParams.AllData, deadlockProtection.Token);
+                paginatedList.Items.Reverse();
+                Assert.Equivalent(ApplyFilter(insertionList, filter), paginatedList.Items);
+            }
+        }
+
+        [Fact]
+        public async Task GetCalculationsListWithPaginationTest()
+        {
+            CancellationTokenSource deadlockProtection = new CancellationTokenSource();
+            deadlockProtection.CancelAfter(TimeSpan.FromSeconds(60));
+
+            using var tempDirHolder = new TempDirectoryHolder();
+            await using var controller = CreateController(tempDirHolder.Directory);
+            await controller.Init(deadlockProtection.Token);
+
+            List<Calculation> insertionList = [
+                CreateCalculationPending(user: "abc"),
+                CreateCalculationInProgress(user: "bcd", timeOffset: TimeSpan.FromMilliseconds(10)),
+                CreateCalculationSuccess(user: "bcd", timeOffset: TimeSpan.FromMilliseconds(20)),
+                CreateCalculationFailed(user: "abc", expression: "log(10) / cos(1)", timeOffset: TimeSpan.FromMilliseconds(30)),
+                CreateCalculationCancelled(user: "abc", timeOffset: TimeSpan.FromMilliseconds(40))
+            ];
+
+            foreach (var item in insertionList)
+            {
+                await controller.AddCalculationAsync(item, deadlockProtection.Token);
+            }
+
+            var paginatedList = await controller.GetCalculationsListAsync(CalculationFilters.NoFilters, new PaginationParams(0, 2), deadlockProtection.Token);
+            Assert.Equal(0u, paginatedList.Offset);
+            Assert.Equal(2u, paginatedList.Limit);
+            Assert.Equal(2, paginatedList.Items.Count);
+            Assert.Null(paginatedList.TotalItemsCount);
+            Assert.Equal(insertionList[insertionList.Count - 1].Id, paginatedList.Items[0].Id);
+            Assert.Equal(insertionList[insertionList.Count - 2].Id, paginatedList.Items[1].Id);
+
+            paginatedList = await controller.GetCalculationsListAsync(CalculationFilters.NoFilters, new PaginationParams(2, 2, true), deadlockProtection.Token);
+            Assert.Equal(2u, paginatedList.Offset);
+            Assert.Equal(2u, paginatedList.Limit);
+            Assert.Equal(2, paginatedList.Items.Count);
+            Assert.NotNull(paginatedList.TotalItemsCount);
+            Assert.Equal(5u, paginatedList.TotalItemsCount.Value);
+            Assert.Equal(insertionList[insertionList.Count - 3].Id, paginatedList.Items[0].Id);
+            Assert.Equal(insertionList[insertionList.Count - 4].Id, paginatedList.Items[1].Id);
         }
 
 
@@ -270,7 +378,7 @@ namespace ExprCalc.Storage.Tests.DatabaseManagement
             int reads = 0;
             while (!bcgTask.IsCompleted)
             {
-                await controller.GetCalculationsListAsync(deadlockProtection.Token);
+                await controller.GetCalculationsListAsync(CalculationFilters.NoFilters, PaginationParams.AllData, deadlockProtection.Token);
                 reads++;
             }
 
@@ -278,8 +386,8 @@ namespace ExprCalc.Storage.Tests.DatabaseManagement
 
             _output.WriteLine($"Reads = {reads}");
 
-            var finalList = await controller.GetCalculationsListAsync(deadlockProtection.Token);
-            Assert.Equal(100, finalList.Count);
+            var finalList = await controller.GetCalculationsListAsync(CalculationFilters.NoFilters, PaginationParams.AllData, deadlockProtection.Token);
+            Assert.Equal(100, finalList.Items.Count);
         }
     }
 }